+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * Remove 'Manage' button from landing page (UIDATIMP-652)
 * Refactor `bigtest/mirage` with `miragejs`.
 * MARC Bib field mapping profile: details for Update-Overrides on Create/Edit screen (UIDATIMP-631)
+* MARC Bib field mapping profile: details for Update-Overrides on View screen (UIDATIMP-632)
 
 ### Bugs fixed:
 * Fix rendering qualifier sections with old data in match profiles details (UIDATIMP-481)